+++ oneflow/operator/copy_op.cpp
@@ -9,7 +9,7 @@ void CopyOp::Init(const OperatorConf& op_conf) {
   auto cnf = new CopyOpConf(op_conf.copy_op_conf());
   mut_pb_op_conf().reset(cnf);
 
-  for (int32_t i = 0; i < cnf->copied_lbns_size(); ++i) {
+  for (int64_t i = 0; i < cnf->copied_lbns_size(); ++i) {
     std::string ibn = "in_" + std::to_string(i);
     EnrollInputBn(ibn);
     CHECK(ibn2lbn_.emplace(ibn, cnf->copied_lbns(i)).second);
